requirements.txt
google-genai
python-dotenv
pymupdf4llm